Add Logentries LogSet Data Source
We'd like to be able to refer to existing LogSets.

This is useful if you want to create a Log under already existing
LogSet.  In order to do this, you need to provide LogSet ID to
Logentries Log resource, but for existing LogSets there's no way to do
this without recreating LogSet (which we don't want).

This data source uses Logentries client library to perform a search by
name and return Terraform data source which you can refer to and query
id and name.

Included tests and documentation